requirements.txt
google-cloud-storage==2.16.0
requests==2.32.2
